requirement.txt
Django>=4.2
djangorestframework>=3.14
python-dotenv>=1.0
psycopg2-binary>=2.9    # or mysqlclient if using MySQL